+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 4.1 (unreleased)
 ----------------
 
+- Fix tests to run with `Products.GenericSetup >= 1.10`.
 
 4.0 (2017-09-14)
 ----------------